+++ src/lib/search.h
@@ -31,5 +31,6 @@ void free_search_expression_list(struct t_list *expr_list);
 bool search_expression_song(const struct mpd_song *song, const struct t_list *expr_list, const struct t_mympd_mpd_tags *any_tag_types);
 bool search_expression_album(const struct t_album *album, const struct t_list *expr_list, const struct t_mympd_mpd_tags *any_tag_types);
 bool search_expression_webradio(const struct t_webradio_data *webradio, const struct t_list *expr_list, const struct t_webradio_tags *any_tag_types);
+bool mympd_search_fuzzy_match(const char *haystack, const char *needle);
 
 #endif